fix: 채팅방 생성 기능 수정

+++ src/main/java/kit/project/whatshouldweeattoday/domain/entity/Restaurant.java
@@ -39,7 +39,7 @@ public class Restaurant {
     // null값을 허용하기 위해
     private Integer pathTime;
     private Double distance;
-    private int count; // 음식종류별 순위
+    private Long count; // 음식종류별 순위
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "bookmark_id")
